Relocate acceptance tests to src/main from src/test
The acceptance tests are now building as an executable uberjar artifact
and not executing during the regular `mvn clean install` process. This
allows the first stage of the pipeline to focus on unit testing and
producing executable jars as artifacts: service and acceptance tests.
The acceptance tests will be run in a later stage of the pipeline. They
can be downloaded as an artifact and executed on any environment with a
JRE.

Typically when acceptance tests fail in an automated qa environment, the
first course of action is to determine if the failures are environment-
specific. By being able to download the exact jar that ran and run it
against another environment, we can efficiently determine if the
environment plays a role in the failure.