Add CI package, acceptance targets
The acceptance target takes a server artifact and an acceptance test
artifact. It starts the server up in a parallel daemon thread and
attempts to ping the server for up to 30 seconds. Once the server comes
online, the acceptance test jar is executed. Once the tests have
completed, the daemon thread containing the server is terminated.

Service and admin service ports are parameterized.

The following conditions fail the build:
* server unreachable in 30 seconds
* acceptance test jar fails to start
* any acceptance test fails